requirements.txt
beautifulsoup4
requests
pandas
numpy
selenium
webdriver-manager
python-dotenv
lxml
html5lib
gspread
google-auth-oauthlib